+++ addons/main/script_component.hpp
@@ -1,0 +1,17 @@
+#define COMPONENT main
+#include "\z\wb_vanilla\addons\main\script_mod.hpp"
+
+// #define DEBUG_MODE_FULL
+// #define DISABLE_COMPILE_CACHE
+// #define CBA_DEBUG_SYNCHRONOUS
+// #define ENABLE_PERFORMANCE_COUNTERS
+
+#ifdef DEBUG_ENABLED_WB_VANILLA
+    #define DEBUG_MODE_FULL
+#endif
+
+#ifdef DEBUG_SETTINGS_WB_VANILLA
+    #define DEBUG_SETTINGS DEBUG_SETTINGS_WB_VANILLA
+#endif
+
+#include "\x\cba\addons\main\script_macros_common.hpp"